move breakpoint definitions/types to Breakpoint utils
