+++ src/utils/ids-breakpoint-utils/ids-breakpoint-utils.ts
@@ -1,4 +1,13 @@
-import { Breakpoints, breakpoints } from '../../core/ids-attributes';
+export type Breakpoints = { [index: string]: string };
+
+export const breakpoints: Breakpoints = {
+  xxl: '1440px',
+  xl: '1280px',
+  lg: '1024px',
+  md: '840px',
+  sm: '600px',
+  xs: '360px'
+};
 
 /**
  * Check for max width media query.
